🏗️ Move components related to post feature into new dedicated folder
