+++ src/features/activity-feed/post/components/Feed/index.jsx
@@ -2,12 +2,12 @@ import PropTypes from 'prop-types'
 import { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { Box, Typography, Avatar, Grid, CircularProgress } from '@mui/material'
-import { getUser, getUserOrganizationName } from '../../features/user/store/userSelectors'
-import { getPosts, getHasMorePosts, getPostLoading } from '../../redux/selectors/feed'
-import { fetchPosts } from '../../redux/thunks/feed'
-import { cleanFeedState } from '../../redux/reducers/feed'
-import { SelectedUserCard } from '../Cards/SelectedUserCard'
-import { PostForm } from '../Forms/PostForm'
+import { getUser, getUserOrganizationName } from '../../../../user/store/userSelectors'
+import { getPosts, getHasMorePosts, getPostLoading } from '../../../../../redux/selectors/feed'
+import { fetchPosts } from '../../../../../redux/thunks/feed'
+import { cleanFeedState } from '../../../../../redux/reducers/feed'
+import { SelectedUserCard } from '../SelectedUserCard'
+import { PostForm } from '../PostForm'
 import { Post } from '../Post'
 import { FeedPlaceholder } from '../FeedPlaceholder'
