+++ src/features/activity-feed/post/components/Feed/style.scss
@@ -1,4 +1,4 @@
-@use '../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/breakpoints' as b;
 
 .c-feed {
     max-width: 600px;
