+++ src/features/activity-feed/post/components/Post/index.jsx
@@ -7,13 +7,13 @@ import { Card, CardActions, CardHeader, CardContent, CircularProgress, Grid, Typ
 import { styled } from '@mui/material/styles'
 import { HashLink } from 'react-router-hash-link'
 import pluralize from 'pluralize'
-import { getPostComments, getPostReactions } from '../../redux/selectors/feed'
-import { fetchComments } from '../../redux/thunks/feed'
-import { getUser } from '../../features/user/store/userSelectors'
-import { CommentForm } from '../Forms/CommentForm'
-import { Comment } from '../Comment'
-import { ReactionButton } from '../Buttons/ReactionButton'
-import { PostReactionsCounter } from '../PostReactionsCounter'
+import { getPostComments, getPostReactions } from '../../../../../redux/selectors/feed'
+import { fetchComments } from '../../../../../redux/thunks/feed'
+import { getUser } from '../../../../user/store/userSelectors'
+import { CommentForm } from '../../../../../components/Forms/CommentForm'
+import { Comment } from '../../../../../components/Comment'
+import { ReactionButton } from '../../../../../components/Buttons/ReactionButton'
+import { PostReactionsCounter } from '../../../../../components/PostReactionsCounter'
 
 import './style.scss'
