+++ src/features/activity-feed/post/components/Post/style.scss
@@ -1,5 +1,5 @@
-@use '../../styles/settings/breakpoints' as b;
-@use '../../styles/settings/colors' as c;
+@use '../../../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/colors' as c;
 
 .c-counter {
     display: flex;
