+++ src/features/user/pages/UserProfile/index.jsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom'
-import { Feed } from '../../../../components/Feed'
+import { Feed } from '../../../activity-feed/post/components/Feed'
 import { ScrollTopButton } from '../../../../components/Buttons/ScrollTopButton'
 import { AuthenticatedLayout } from '../../../../layouts/AuthenticatedLayout'
 import './style.scss'
